update0907

+++ src/main/java/com/spring/domain/BoardVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class BoardVO {
-	private Long bno;
+	private int bno;
 	private String title, content, writer;
 	private Date regdate, updateDate;
 }
